+++ dash/db/sessions.php
@@ -223,12 +223,12 @@ class sessions
 
 			$get  = self::get($get_cookie);
 
-			if(isset($get['ip']) && intval($get['ip']) !== intval(\dash\server::ip(true)))
+			if(isset($get['ip']) && floatval($get['ip']) !== floatval(\dash\server::ip(true)))
 			{
 				\dash\log::set('ipChangeRememberExpire');
 			}
 
-			if(isset($get['agent_id']) && intval($get['agent_id']) !== intval(\dash\agent::get(true)))
+			if(isset($get['agent_id']) && floatval($get['agent_id']) !== floatval(\dash\agent::get(true)))
 			{
 				\dash\log::set('agentChangeRememberExpire');
 			}
@@ -313,13 +313,13 @@ class sessions
 				// check ip and agent and if updated, update it
 				$update_current = [];
 
-				if(isset($get['ip']) && intval($get['ip']) !== intval(\dash\server::ip(true)))
+				if(isset($get['ip']) && floatval($get['ip']) !== floatval(\dash\server::ip(true)))
 				{
 					\dash\log::set('sessionIPupdated', ['code' => isset($get['id']) ? $get['id'] : null]);
 					$update_current['ip'] = \dash\server::ip(true);
 				}
 
-				// if(isset($get['agent_id']) && intval($get['agent_id']) !== intval(\dash\agent::get(true)))
+				// if(isset($get['agent_id']) && floatval($get['agent_id']) !== floatval(\dash\agent::get(true)))
 				// {
 				// 	$update_current['agent_id'] = \dash\agent::get(true);
 				// }
